src/alert_system.rs
use crate::log_processor::LogEntry;

pub struct AlertConfig {
    // TODO
    // Define alert configuration here
    // For example, thresholds, patterns, etc.
}

pub struct AlertSystem {
    config: AlertConfig,
}

impl AlertSystem {
    // Constructor for AlertSystem
    pub fn new(config: AlertConfig) -> Self {
        AlertSystem { config }
    }

    // Method to check for alerts based on log entries
    pub fn check_alerts(&self, logs: &[LogEntry]) {
        for log in logs {
            if self.should_trigger_alert(log) {
                self.trigger_alert(log);
            }
        }
    }

    // Determine if an alert should be triggered
    fn should_trigger_alert(&self, log: &LogEntry) -> bool {
        // TODO
        // Implement logic to determine if an alert should be triggered
        // For example, check for certain patterns, error levels, frequency, etc.

        // This is just a placeholder condition
        log.level == "ERROR"
    }

    // Trigger an alert
    fn trigger_alert(&self, log: &LogEntry) {
        // TODO
        // Implement logic for what happens when an alert is triggered
        // For example, sending an email, pushing a notification, etc.

        println!("Alert triggered: {}", log.message);
    }
}
