tests/alert_system_tests.rs
#[cfg(test)]
mod tests {
    use super::*;
    use crate::log_processor::LogEntry;

    // Helper function to create a sample log entry
    fn create_sample_log_entry(level: &str, message: &str) -> LogEntry {
        LogEntry {
            timestamp: "2023-11-17T12:00:00".to_string(),
            level: level.to_string(),
            message: message.to_string(),
        }
    }

    #[test]
    fn test_should_trigger_alert() {
        let alert_config = AlertConfig {
            // TODO
            // Define the criteria for triggering an alert
            // Example: Alert for ERROR level logs
            // ... (initialize AlertConfig here)
        };
        let alert_system = AlertSystem::new(alert_config);

        let error_log = create_sample_log_entry("ERROR", "Error occurred");
        let info_log = create_sample_log_entry("INFO", "Informational message");

        assert!(alert_system.should_trigger_alert(&error_log));
        assert!(!alert_system.should_trigger_alert(&info_log));
    }

    #[test]
    fn test_check_alerts() {
        let alert_config = AlertConfig {
            // TODO
            // Initialize AlertConfig here
            // ...
        };
        let alert_system = AlertSystem::new(alert_config);

        let logs = vec![
            create_sample_log_entry("ERROR", "First error"),
            create_sample_log_entry("INFO", "Just an info"),
            create_sample_log_entry("ERROR", "Second error"),
        ];

        // Assuming `check_alerts` returns a count of triggered alerts or similar
        let triggered_alerts = alert_system.check_alerts(&logs);

        // Adjust the assertion based on the actual implementation
        assert_eq!(triggered_alerts, 2);
    }

    // Additional tests for other functionalities of `AlertSystem` can be added here
}
